% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naming_cols}
\alias{naming_cols}
\title{This function names each column the right way in the final dataset}
\usage{
naming_cols(df, well_id)
}
\arguments{
\item{df}{This a dataframe/ tibble}

\item{well_id}{This is a character vector of length 1 which will be used to
name well in the final tibble}
}
\value{
A list/ tibble
}
\description{
This function names each column the right way in the final dataset
}
\examples{
df <- tibble::tibble(values = LETTERS[1:10], b = 1:10, c = 21:30)

new_df <- naming_cols(df, "Wells")

head(new_df)
}
