% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plate_params}
\alias{plate_params}
\title{This function generates different plate parameters if the data is formatted
correctly
\code{no_of_plates}: Calculates the number of plates in the raw data
\code{count_rows_theoretical}: Counts the number of rows each plate type should have
for a given the no_of_plates
\code{plate_type}: Guesses the type of microtiter plate type
\code{cols}: Columns vector for each plate type from raw data}
\usage{
plate_params(data, n_cols)
}
\arguments{
\item{data}{This takes the raw data after imported by \code{tidy_plate} or \code{check_plate}
functions}

\item{n_cols}{This takes the number of columns in the data as integer}
}
\value{
A list of several plate parameters
}
\description{
This function generates different plate parameters if the data is formatted
correctly
\code{no_of_plates}: Calculates the number of plates in the raw data
\code{count_rows_theoretical}: Counts the number of rows each plate type should have
for a given the no_of_plates
\code{plate_type}: Guesses the type of microtiter plate type
\code{cols}: Columns vector for each plate type from raw data
}
