% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_plate.R
\name{pull_plate}
\alias{pull_plate}
\title{Subset individual plates from the input file}
\usage{
pull_plate(file, sheet = 1, plate_id)
}
\arguments{
\item{file}{This is the path to a xlsx or csv file containing data for the
following types of plates: 6, 12, 24, 48, 96, 384, and 1536.}

\item{sheet}{If file type is xlsx this is the sheet name (character) or
number (integer).}

\item{plate_id}{Character or numeric vector that will be used to subset the file.}
}
\value{
a list of tibbles
}
\description{
Subset individual plates from the input file
}
\examples{
file_path <- system.file("extdata", "example_12_well.xlsx", package = "tidyplate")

n_id = c(1, 3)
c_id = c("drug", "percent_survived")

data_n <- pull_plate(file = file_path, plate_id = n_id)
data_c <- pull_plate(file = file_path, plate_id = c_id)

print(data_n)
print(data_c)
}
