% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_plate.R
\name{tidy_plate}
\alias{tidy_plate}
\title{Transforms a plate to a tidy dataframe/ tibble}
\usage{
tidy_plate(file, sheet = 1)
}
\arguments{
\item{file}{This is the path to a xlsx or csv file containing data for the
following types of plates: 6, 12, 24, 48, 96, 384, and 1536. The plate format
is described below.}

\item{sheet}{If file type is xlsx this is the sheet name (character) or
number (integer).}
}
\value{
A tidy dataframe/ tibble
}
\description{
Transforms a plate to a tidy dataframe/ tibble
}
