% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_plates.R
\name{view_plates}
\alias{view_plates}
\title{Returns the name of each plate in the file}
\usage{
view_plates(file, sheet = 1)
}
\arguments{
\item{file}{This is the path to a xlsx or csv file containing data for the
following types of plates: 6, 12, 24, 48, 96, 384, and 1536.}

\item{sheet}{If file type is xlsx this is the sheet name (character) or
number (integer).}
}
\value{
A character vector
}
\description{
Returns the name of each plate in the file
}
\examples{
file_path <- system.file("extdata", "example_12_well.xlsx", package = "tidyplate")

data_12 <- view_plates(file = file_path)

data_12
}
