% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plate.R
\name{build_plate}
\alias{build_plate}
\title{Returns a skeleton csv or xlsx file}
\usage{
build_plate(
  plate_type = 6,
  n_plates = 1,
  file_type = c("csv", "xlsx"),
  plate_names = NULL,
  file = NULL
)
}
\arguments{
\item{plate_type}{Type of microwell plate (double)}

\item{n_plates}{The total number of plates (double)}

\item{file_type}{Type of file; either a "csv" (default) or "xlsx"}

\item{plate_names}{A character vector that will be assigned to each plate}

\item{file}{The path where the file will be saved (defaults to current directory)}
}
\value{
A csv or xlsx file
}
\description{
Returns a skeleton csv or xlsx file
}
