% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plate.R
\name{generate_plate}
\alias{generate_plate}
\title{Generates a microwell plate shaped csv or excel files from tibble or dataframe.}
\usage{
generate_plate(x, well_id, plate_type, file_out, file_type = c("csv", "xlsx"))
}
\arguments{
\item{x}{A dataframe or tibble.}

\item{well_id}{A character string or an integer which points to the column
containing the well ids.}

\item{plate_type}{A specific integer (6, 12, 24, 48, 96, 384, 1536)
indicating the type of microwell plate.}

\item{file_out}{A character string with the name (with path) of the output
file name.}

\item{file_type}{A character string with the type of exported file ("csv",
"xlsx"). The default is "csv".}
}
\value{
A csv or xlsx file.
}
\description{
Generates a microwell plate shaped csv or excel files from tibble or dataframe.
}
