% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateTransformR-package.R
\docType{package}
\name{tidyplate-package}
\alias{tidyplate}
\alias{tidyplate-package}
\title{tidyplate: What the Package Does (Title Case)}
\description{
This package helps researchers transform plate data to a tidy dataframe/ tibble.
}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID})

}
\keyword{internal}
