% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateTransformR-package.R
\docType{package}
\name{tidyplate-package}
\alias{tidyplate}
\alias{tidyplate-package}
\title{tidyplate: Transform Microplate Data into Tidy Dataframes}
\description{
The goal of 'tidyplate' is to help researchers convert different types of microplates into tidy dataframes which can be used in data analysis. It accepts xlsx and csv files formatted in a specific way as input. It supports all types of standard microplate formats such as 6-well, 12-well, 24-well, 48-well, 96-well, 384-well, and, 1536-well plates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shubhamdutta26/tidyplate}
  \item \url{https://www.shubhamdutta.com/tidyplate/}
  \item Report bugs at \url{https://github.com/shubhamdutta26/tidyplate/issues}
}

}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID}) [copyright holder]

}
\keyword{internal}
